+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/Region.cs
@@ -1,11 +1,5 @@
-// 
-// Generated code
-// 
-
 using System;
 using System.Collections.Generic;
-using Microsoft.Data.Entity;
-using Microsoft.Data.Entity.Metadata;
 
 namespace Kendo.Mvc.Examples.Models
 {
@@ -15,12 +9,10 @@ namespace Kendo.Mvc.Examples.Models
         {
             Territories = new HashSet<Territory>();
         }
-        
-        // Properties
-        public int RegionID { get; set; }
+
+        public long RegionID { get; set; }
         public string RegionDescription { get; set; }
-        
-        // Navigation Properties
+
         public virtual ICollection<Territory> Territories { get; set; }
     }
 }
