+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/Shipper.cs
@@ -1,11 +1,5 @@
-// 
-// Generated code
-// 
-
 using System;
 using System.Collections.Generic;
-using Microsoft.Data.Entity;
-using Microsoft.Data.Entity.Metadata;
 
 namespace Kendo.Mvc.Examples.Models
 {
@@ -15,13 +9,11 @@ namespace Kendo.Mvc.Examples.Models
         {
             Orders = new HashSet<Order>();
         }
-        
-        // Properties
-        public int ShipperID { get; set; }
+
+        public long ShipperID { get; set; }
         public string CompanyName { get; set; }
         public string Phone { get; set; }
-        
-        // Navigation Properties
+
         public virtual ICollection<Order> Orders { get; set; }
     }
 }
